Splits reducer and adds selectors
